requirements.txt
tensorflow>=2.8.0
numpy>=1.21.0
pandas>=1.3.0
cryptography>=3.4.7
pyyaml>=5.4.1
requests>=2.26.0
python-logging>=0.4.9.6
ipfshttpclient>=0.8.0
tenseal>=0.3.0
differential-privacy>=0.2.0 